requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
asttokens
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama
comm
datasets==2.18.0
debugpy
decorator
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
exceptiongroup
executing
filelock==3.13.3
Flask==3.0.2
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.42
happytransformer==3.0.0
huggingface-hub==0.26.1
idna==3.6
importlib_metadata==7.1.0
intel-openmp==2021.4.0
ipykernel
ipython
itsdangerous==2.1.2
jedi
Jinja2==3.1.3
joblib==1.4.2
jupyter_client
jupyter_core
MarkupSafe==2.1.5
matplotlib-inline
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso
pickleshare
pillow==11.0.0
platformdirs
prompt_toolkit
protobuf==4.25.3
psutil==5.9.8
pure_eval
pyarrow==15.0.2
pyarrow-hotfix==0.6
Pygments
pymongo==4.6.3
PySocks==1.7.1
python-dateutil
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.5.2
scipy==1.12.0
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==1.44.0
setproctitle==1.3.3
six
smmap==5.0.1
soupsieve==2.5
stack-data
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.3.0
tornado
tqdm==4.66.2
traitlets
transformers==4.46.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wandb==0.16.5
wcwidth
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1